requirements.txt
openai
apify-client
python-dotenv
pandas
matplotlib
